fix to verify correct verse in updateVerseEditStatesAndCheckAlignments()
