+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "checking-tool-wrapper",
   "description": "Checking tool wrapper for translationCore App",
-  "version": "3.0.2",
+  "version": "3.0.2-beta.3",
   "main": "dist/bundle.js",
   "scripts": {
     "lint": "eslint ./src",
